+++ src/lifobuffer.c
@@ -1,8 +1,9 @@
 #include "lifobuffer.h"
 
 #include <string.h>
+#include <stdint.h>
 
-void LifoBuffer_Init(LifoBuffer* lifo, void* buff, uint32_t elementSize, uint32_t numElements)
+void LifoBuffer_Init(LifoBuffer* lifo, void* buff, size_t elementSize, size_t numElements)
 {
 	lifo->Buffer	  = buff;
 	lifo->ElementSize = elementSize;
@@ -17,7 +18,7 @@ void* LifoBuffer_Add(LifoBuffer* lifo)
 	if (lifo->Full)
 		return NULL;
 
-	uint32_t offset = lifo->Head * lifo->ElementSize;
+	size_t offset = lifo->Head * lifo->ElementSize;
 
 	lifo->Full	= ++lifo->Head == lifo->NumElements;
 	lifo->Empty = false; // Cannot be full - just added element
@@ -30,7 +31,7 @@ void* LifoBuffer_Remove(LifoBuffer* lifo)
 	if (lifo->Empty)
 		return NULL;
 
-	uint32_t offset = --lifo->Head * lifo->ElementSize;
+	size_t offset = --lifo->Head * lifo->ElementSize;
 
 	lifo->Empty = lifo->Head == 0;
 	lifo->Full	= false; // Cannot be empty - just removed element
